Fix topping stock validation and sales flow

+++ src/main/java/com/deli/repository/SaleRepository.java
@@ -17,7 +17,7 @@ public interface SaleRepository extends JpaRepository<Sale, Long> {
     List<Sale> findBySaleDateBetweenOrderByCreatedAtDesc(LocalDate start, LocalDate end);
 
     @Query("select coalesce(sum(s.quantity), 0) from Sale s where s.saleDate = :date")
-    Integer totalUnitsByDate(LocalDate date);
+    Long totalUnitsByDate(LocalDate date);
 
     @Query("select coalesce(sum(s.total), 0) from Sale s where s.saleDate = :date")
     java.math.BigDecimal totalAmountByDate(LocalDate date);
